% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cibersort.r
\name{get_cibersort}
\alias{get_cibersort}
\title{character the fractions of tumor-infiltrating immune cells}
\usage{
get_cibersort(user_file, perm = 1000, QN = TRUE)
}
\arguments{
\item{user_file}{the path of gene expression data.frame.}

\item{perm}{Number of replacements.}

\item{QN}{Quantile normalization}
}
\value{
Impute a data.frame of Cell fractions.
}
\description{
character the fractions of tumor-infiltrating immune cells
}
\examples{
# user_file = 'path/Cibersort_ucs.txt'
# result <- get_cibersort(user_file,perm = 1000, QN = TRUE)
}
