% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImTH.r
\name{ImTH}
\alias{ImTH}
\title{quantifying intratumor immune heterogeneity}
\usage{
ImTH(data, type = c("CIBERSORT", "custom"))
}
\arguments{
\item{data}{A data.frame of proportions of immune cell types across samples.}

\item{type}{A character vector indicating the type of "CIBERSORT", "custom".}
}
\value{
A data frame containing sample IDs and intratumor immune heterogeneity scores -- ImTH score
}
\description{
quantifying intratumor immune heterogeneity
}
\examples{
# Example usage:
# result_bulk <- ImTH(data, type = "CIBERSORT")
# result_sc <- ImTH(data, type = "custom")
}
