% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImTH.r
\name{ImTH}
\alias{ImTH}
\title{quantifying Immune Tumor Heterogeneity}
\usage{
ImTH(data, type = c("CIBERSORT", "custom"))
}
\arguments{
\item{data}{A data.frame of the proportion of the immune cell across samples.}

\item{type}{A character vector indicating the type of "CIBERSORT", "custom".}
}
\value{
A data frame containing sample IDs and Immune Tumor Heterogeneity scores -- ImTH score
}
\description{
quantifying Immune Tumor Heterogeneity
}
\examples{
# Example usage:
# result_bulk <- ImTH(data, type = "CIBERSORT")
# result_sc <- ImTH(data, type = "custom")
}
